add start gunicorn server to cli command
