+++ setup.py
@@ -24,8 +24,7 @@ install_requires = [
     'Werkzeug',
     'pathlib2',
     'requests',
-    'packaging',
-    'gunicorn'
+    'packaging'
 ]
 
 dev_requires = [
@@ -43,8 +42,9 @@ dev_requires = [
 cv2 = [ 'opencv-python' ]
 pytorch  = [ 'torch', 'torchvision' ]
 tensorflow = [ 'tensorflow' ]
+gunicorn = [ 'gunicorn' ]
 
-optional_requires = cv2 + pytorch + tensorflow
+optional_requires = cv2 + pytorch + tensorflow + gunicorn
 dev_all = install_requires + dev_requires + optional_requires
 
 tests_require = [
